cyberrecon/requirements.txt
Flask==3.0.3
requests==2.32.3
beautifulsoup4==4.12.3
python-whois==0.9.4
dnspython==2.6.1
playwright==1.48.0
aiofiles==24.1.0
reportlab==4.2.2
python-dotenv==1.0.1
